Move more functionality to the AbstractDateCalculator. Got a bit stuck on WorkingWeek, may be that could be <parameterised>...

+++ datecalc-joda/src/main/java/net/objectlab/kit/datecalc/joda/DefaultPeriodCountCalculator.java
@@ -82,7 +82,7 @@ public class DefaultPeriodCountCalculator implements PeriodCountCalculator {
 
                 final int diff1 = new Period(start, endOfStartYear, PeriodType.days()).getDays();
                 final int diff2 = new Period(startOfEndYear, end, PeriodType.days()).getDays();
-                diff = (diff1 + 1.0) / (double) start.dayOfYear().getMaximumValue() + (endYear - startYear - 1.0) + (diff2)
+                diff = (diff1 + 1.0) / start.dayOfYear().getMaximumValue() + (endYear - startYear - 1.0) + (diff2)
                         / (double) end.dayOfYear().getMaximumValue();
             }
